+++ xraise.h
@@ -23,7 +23,6 @@
 #include <string.h>
 #include <stdarg.h>
 #include <time.h>
-
 #ifdef _countof
 #define CountItems(a) ((int)_countof(a))
 #else
@@ -38,7 +37,7 @@ extern bool noisy;		// Used to key debugging logs
 		Utility routines
 */
 bool frontcmp( const char* pattern, const char* s);
-void EndianSwap( void* thing, int size);
+void EndianSwap( void* thing, size_t size);
 char tolower_c( char ch);
 char toupper_c( char ch);
 int strli( const char* s);		// Integer version of strlen
